Finding error

+++ libraries/includes/vector.h
@@ -6,7 +6,7 @@
 /*   By: shillebr <shillebr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/27 06:26:00 by xrhoda            #+#    #+#             */
-/*   Updated: 2018/09/27 16:28:42 by shillebr         ###   ########.fr       */
+/*   Updated: 2018/09/27 19:03:34 by shillebr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@
 typedef struct	s_vector
 {
 	void		**items;
-	double		capacity;
-	double		total;
+	int			capacity;
+	int			total;
 }				t_vector;
 
 void			vector_init(t_vector *v);
